fix wrong value again

+++ exercises/07_loop/main.cpp
@@ -12,7 +12,7 @@ static unsigned long long fibonacci(int i) {
     cache[1] = 1;
     cached = 2;
     // TODO: 设置正确的循环条件
-    for (cached = 2; cached <=i; ++cached) {
+    for (cached = 2; cached <=(unsigned long long)i; ++cached) {
         cache[cached] = cache[cached - 1] + cache[cached - 2];
     }
     return cache[i];
